bug fixes, clean up, make implementation more functional
